# backend/src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Blockchain Identity System API",
    "description": "RESTful API for managing and verifying blockchain-based digital identities",
    "version": "1.0.0",
    "contact": {
      "email": "support@example.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.identity.example.com",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/api/identity/{address}": {
      "get": {
        "tags": ["Identity"],
        "summary": "Get identity details",
        "description": "Retrieve identity details for a specific Ethereum address",
        "operationId": "getIdentity",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Ethereum address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Identity found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Identity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/reputation/{address}": {
      "get": {
        "tags": ["Reputation"],
        "summary": "Get reputation",
        "description": "Get reputation score for a specific address",
        "operationId": "getReputation",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Ethereum address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Reputation retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "address": {
                          "type": "string",
                          "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
                        },
                        "score": {
                          "type": "integer",
                          "example": 100
                        },
                        "isBanned": {
                          "type": "boolean",
                          "example": false
                        },
                        "lastUpdate": {
                          "type": "string",
                          "format": "date-time",
                          "example": "2023-03-15T14:30:00Z"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Reputation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/gateway/identity/{address}": {
      "get": {
        "tags": ["Gateway"],
        "summary": "Get identity via gateway",
        "description": "Gateway endpoint to retrieve identity information for third-party integration",
        "operationId": "gatewayGetIdentity",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Ethereum address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Identity found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayIdentityResponse"
                }
              }
            }
          },
          "404": {
            "description": "Identity not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - API key invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gateway/reputation/{address}": {
      "get": {
        "tags": ["Gateway"],
        "summary": "Get reputation via gateway",
        "description": "Gateway endpoint to retrieve reputation information for third-party integration",
        "operationId": "gatewayGetReputation",
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "description": "Ethereum address",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Reputation found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayReputationResponse"
                }
              }
            }
          },
          "404": {
            "description": "Reputation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - API key invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/gateway/authenticate": {
      "post": {
        "tags": ["Gateway"],
        "summary": "Authenticate via gateway",
        "description": "Gateway endpoint for third-party authentication",
        "operationId": "gatewayAuthenticate",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GatewayAuthRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayAuthResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "IdentityResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "tokenId": {
                "type": "string",
                "description": "NFT token ID",
                "example": "1"
              },
              "owner": {
                "type": "string",
                "description": "Ethereum address of the owner",
                "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
              },
              "did": {
                "type": "string",
                "description": "Decentralized identifier",
                "example": "did:eth:0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
              },
              "creationDate": {
                "type": "string",
                "description": "Unix timestamp of creation",
                "example": "1678897200"
              },
              "lastUpdate": {
                "type": "string",
                "description": "Unix timestamp of last update",
                "example": "1678983600"
              },
              "verificationLevel": {
                "type": "integer",
                "description": "Verification level (0-3)",
                "example": 2
              },
              "isRecoverable": {
                "type": "boolean",
                "description": "Whether the identity is recoverable",
                "example": true
              },
              "recoveryAddress": {
                "type": "string",
                "description": "Recovery address if set",
                "example": "0x123ABC..."
              },
              "lastVerificationDate": {
                "type": "string",
                "description": "Unix timestamp of last verification",
                "example": "1678983600"
              }
            }
          }
        }
      },
...
      "GatewayIdentityResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "address": {
                "type": "string",
                "description": "Ethereum address",
                "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
              },
              "did": {
                "type": "string",
                "description": "Decentralized identifier",
                "example": "did:eth:0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
              },
              "verificationLevel": {
                "type": "integer",
                "description": "Verification level (0-3)",
                "example": 2
              },
              "creationDate": {
                "type": "string",
                "format": "date-time",
                "description": "Creation date",
                "example": "2023-03-15T14:30:00Z"
              },
              "isVerified": {
                "type": "boolean",
                "description": "Whether the identity is verified",
                "example": true
              },
              "recoverable": {
                "type": "boolean",
                "description": "Whether the identity is recoverable",
                "example": true
              }
            }
          }
        }
      },
      "GatewayReputationResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "address": {
                "type": "string",
                "description": "Ethereum address",
                "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
              },
              "score": {
                "type": "integer",
                "description": "Reputation score",
                "example": 100
              },
              "lastUpdate": {
                "type": "string",
                "format": "date-time",
                "description": "Last update time",
                "example": "2023-03-15T14:30:00Z"
              },
              "history": {
                "type": "array",
                "description": "Score history",
                "items": {
                  "type": "object",
                  "properties": {
                    "timestamp": {
                      "type": "string",
                      "format": "date-time",
                      "example": "2023-03-15T14:30:00Z"
                    },
                    "score": {
                      "type": "integer",
                      "example": 100
                    },
                    "reason": {
                      "type": "string",
                      "example": "Initial score"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "GatewayAuthRequest": {
        "type": "object",
        "required": ["address", "signature", "message", "timestamp"],
        "properties": {
          "address": {
            "type": "string",
            "description": "Ethereum address",
            "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
          },
          "signature": {
            "type": "string",
            "description": "Ethereum signature",
            "example": "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef1b"
          },
          "message": {
            "type": "string",
            "description": "Message that was signed",
            "example": "Authenticate to Identity System: 1678897200"
          },
          "timestamp": {
            "type": "integer",
            "description": "Unix timestamp (within 5 minutes)",
            "example": 1678897200
          }
        }
      },
      "GatewayAuthResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "address": {
                "type": "string",
                "description": "Ethereum address",
                "example": "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
              },
              "authenticated": {
                "type": "boolean",
                "description": "Authentication result",
                "example": true
              },
              "identity": {
                "type": "object",
                "properties": {
                  "did": {
                    "type": "string",
                    "example": "did:eth:0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
                  },
                  "verificationLevel": {
                    "type": "integer",
                    "example": 2
                  },
                  "isVerified": {
                    "type": "boolean",
                    "example": true
                  }
                }
              },
              "reputation": {
                "type": "object",
                "properties": {
                  "score": {
                    "type": "integer",
                    "example": 100
                  }
                }
              },
              "timestamp": {
                "type": "integer",
                "description": "Unix timestamp",
                "example": 1678897200
              }
            }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "description": "Error message",
            "example": "Invalid signature"
          },
          "code": {
            "type": "string",
            "description": "Error code (optional)",
            "example": "INVALID_SIGNATURE"
          }
        }
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Authorization header using the Bearer scheme"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key for gateway access"
      }
    }
  }
}